Move ID generation into their own classes
