+++ src/blatternfly/Hosting/WebAssemblyHostBuilderExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Blatternfly.Components;
 using Blatternfly.Interop;
+using Blatternfly.Utilities;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.AspNetCore.Components.WebAssembly.Hosting;
@@ -9,6 +10,8 @@ public static class WebAssemblyHostBuilderExtensions
     public static WebAssemblyHostBuilder AddBlatternfly(this WebAssemblyHostBuilder builder)
     {
         // Singleton services
+        builder.Services.AddSingleton<ISequentialIdGenerator, SequentialIdGenerator>();
+        builder.Services.AddSingleton<IRandomIdGenerator, RandomIdGenerator>();
         builder.Services.AddSingleton<IDomUtils, DomUtils>();
         builder.Services.AddSingleton<IWindowObserver, WindowObserver>();
         builder.Services.AddSingleton<IPortalConnector, PortalConnector>();
