+++ src/blatternfly/Utilities/IRandomIdGenerator.cs
@@ -1,0 +1,6 @@
+namespace Blatternfly.Utilities;
+
+public interface IRandomIdGenerator
+{
+    string GenerateId(string prefix = "pf");
+}
