+++ src/blatternfly/Utilities/ISequentialIdGenerator.cs
@@ -1,0 +1,6 @@
+namespace Blatternfly.Utilities;
+
+public interface ISequentialIdGenerator
+{
+    string GenerateId(string prefix = "pf");
+}
